+++ Backend/PayAdjustments/PayAdjustmentsConfig.js
@@ -14,7 +14,7 @@
  *******************************************************/
 
 const PayTrackerPayAdjustmentsConfig = Object.freeze({
-  VERSION: '3.0.5',
+  VERSION: '3.0.6',
 
   SHEETS: Object.freeze({
     ADJUSTMENTS: Object.freeze({